style: move discord-related files to own package
