--- core/src/main/java/st/networkers/discordbooks/channel/Receivable.java
@@ -1,8 +1,0 @@
-package st.networkers.discordbooks.channel;
-
-import st.networkers.discordbooks.message.Sendable;
-
-public interface Receivable {
-    Receivable receive(Sendable message);
-
-}
