+++ core/src/main/java/st/networkers/discordbooks/discord/channel/Receivable.java
@@ -1,0 +1,8 @@
+package st.networkers.discordbooks.discord.channel;
+
+import st.networkers.discordbooks.discord.message.Sendable;
+
+public interface Receivable {
+    Receivable receive(Sendable message);
+
+}
